Reference

Contents

Reference

Contents

Index

Index

GraphDynamicalSystems.QN

GraphDynamicalSystems.QualitativeNetwork

GraphDynamicalSystems.QualitativeNetwork

DynamicalSystemsBase.current_state

GraphDynamicalSystems.entities

GraphDynamicalSystems.get_graph

GraphDynamicalSystems.get_n_entities

GraphDynamicalSystems.get_schedule

GraphDynamicalSystems.QN — Type

struct QualitativeNetwork{N_Entities, Schedule, Graph<:MetaGraphsNext.MetaGraph} <: GraphDynamicalSystems.GraphDynamicalSystem{N_Entities, Schedule}

Alias for a QualitativeNetwork.

GraphDynamicalSystems.QualitativeNetwork — Type

struct QualitativeNetwork{N_Entities, Schedule, Graph<:MetaGraphsNext.MetaGraph} <: GraphDynamicalSystems.GraphDynamicalSystem{N_Entities, Schedule}

A graph dynamical system with a finite domain. State values of each entity are limited to change by at most 1 per time step.

GraphDynamicalSystems.QualitativeNetwork — Method

QualitativeNetwork(
    update_functions::AbstractDict{E, F},
    domains
) -> GraphDynamicalSystems.QualitativeNetwork{_A, GraphDynamicalSystems.Synchronous(), G} where {_A, G<:(MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, _A, _B, _C, Nothing, MetaGraphsNext.var"#MetaGraph##14#MetaGraph##15", Float64} where {_A, _B, _C})}

Create a QualitativeNetwork from the dictionary update_functions which should map from entities (E) to their functions (F)

The entity names (E in the signature) can be anything, while the functions (F) are required to either

be a numerical constant (like 1), or a reference to a single entity (like A)

implement the TermInterface.jl interface. Any terminal nodes in the functions must be numerical constants or reference an entity.

DynamicalSystemsBase.current_state — Method

current_state(
    gds::GraphDynamicalSystems.GraphDynamicalSystem
) -> Any

Get the state of the GDS.

GraphDynamicalSystems.entities — Method

entities(
    gds::GraphDynamicalSystems.GraphDynamicalSystem
) -> Any

List all entities in gds.

GraphDynamicalSystems.get_graph — Method

get_graph(
    gds::GraphDynamicalSystems.GraphDynamicalSystem
) -> Any

Get the underlying graph of the GDS.

GraphDynamicalSystems.get_n_entities — Method

get_n_entities(
    _::GraphDynamicalSystems.GraphDynamicalSystem{N, S}
) -> Any

Get the number of entities N in the GDS.

GraphDynamicalSystems.get_schedule — Method

get_schedule(
    _::GraphDynamicalSystems.GraphDynamicalSystem{N, S}
) -> Any

Get the schedule for the GDS.