Reference

Contents

Reference

Contents

Index

Index

GraphDynamicalSystems.Schedule.Asynchronous

GraphDynamicalSystems.Schedule.Determinism

GraphDynamicalSystems.Schedule.ScheduleStyle

GraphDynamicalSystems.Schedule.Synchronous

GraphDynamicalSystems.Constructors.vertex_function_to_edgelist

GraphDynamicalSystems.Conversions.to_arbitrary_steppable

GraphDynamicalSystems.Conversions.to_discrete_system

GraphDynamicalSystems.Conversions.to_simple_graph

GraphDynamicalSystems.Interface.domain

GraphDynamicalSystems.Interface.edges

GraphDynamicalSystems.Interface.get_state

GraphDynamicalSystems.Interface.schedule_style

GraphDynamicalSystems.Interface.set_state!

GraphDynamicalSystems.Interface.update_function

GraphDynamicalSystems.Interface.vertices

GraphDynamicalSystems.Schedule.is_deterministic

GraphDynamicalSystems.Schedule.is_deterministic

GraphDynamicalSystems.Schedule.to_update

GraphDynamicalSystems.Interface.domain — Function

domain(gds)
domain(gds, vertices)

Get the domain of (all/some) vertices of the gds.

GraphDynamicalSystems.Interface.edges — Function

edges(gds)

Get the edges of the gds.

GraphDynamicalSystems.Interface.get_state — Function

get_state(gds)
get_state(gds, vertices)

Get the state of (all/some) vertices of the gds.

GraphDynamicalSystems.Interface.schedule_style — Function

schedule_style(gds)

Get the schedule style of the gds.

Also referred to as the "update scheme".

GraphDynamicalSystems.Interface.set_state! — Function

set_state!(gds, state)
set_state!(gds, vertex_state_pairs)

Set the state of the entire gds or a subset of vertices' states.

GraphDynamicalSystems.Interface.update_function — Function

update_function(gds)
update_function(gds, vertices)

Get the update function of (all/some) vertices of the gds.

Also referred to as "vertex function(s)".

GraphDynamicalSystems.Interface.vertices — Function

vertices(gds)

Get the vertices of the gds.

GraphDynamicalSystems.Schedule.Asynchronous — Type

Asynchronous

An asynchronous schedule randomly chooses a vertex to update at each timestep.

GraphDynamicalSystems.Schedule.Determinism — Type

Determinism

Trait defining the determinism of a schedule style.

GraphDynamicalSystems.Schedule.ScheduleStyle — Type

ScheduleStyle

Trait defining the schedule of a GDS.

A schedule defines which vertices are updated and in which order. Two basic common schedule types are Asynchronous and Synchronous.

A new schedule type MySchedule<:ScheduleStyle should implement to_update(::MySchedule, gds) and determinism(::MySchedule).

GraphDynamicalSystems.Schedule.Synchronous — Type

Synchronous

A synchronous schedule updates all vertices at each timestep.

GraphDynamicalSystems.Schedule.is_deterministic — Method

is_deterministic(gds)

Whether the gds has a deterministic schedule.

GraphDynamicalSystems.Schedule.is_deterministic — Method

is_deterministic(s::ScheduleStyle)

Whether schedule s is deterministic.

GraphDynamicalSystems.Schedule.to_update — Method

to_update(gds)

Get the vertices of the gds to update at the current timestep.

Implement to_update(::ScheduleStyle, gds) for new schedule types.

GraphDynamicalSystems.Conversions.to_arbitrary_steppable — Function

to_arbitrary_steppable(gds)

Convert gds to a DynamicalSystemsBase.ArbitrarySteppable.

gds is required to implement the GDS interface.

Tip

Use this method when an object <:DynamicalSystem is necessary. The DynamicalSystems ecosystem currently requires this for most (all?) of its functionality. Until a trait-based version of the ecosystem arrives, then one must use this method (or one like it) to construct something that is <:DynamicalSystem.

Note

Requires DynamicalSystemsBase to be loaded.

GraphDynamicalSystems.Conversions.to_discrete_system — Function

to_discrete_system(gds)

Convert gds to a AlgebraicDynamics.DWDDynam.DiscreteMachine.

gds is required to implement the GDS interface.

Note

Currently unimplemented as a number of dependencies within the AlgebraicJulia ecosystem are not up to date and conflict with those of other extensions.

GraphDynamicalSystems.Conversions.to_simple_graph — Function

to_simple_graph(gds)

Convert gds to a Graphs.SimpleDiGraph.

gds is required to implement the GDS interface.

Tip

Use this method when an object <:AbstractGraph is necessary. The Graphs ecosystem currently requires this for most (all?) of its functionality. Until a trait-based version of the ecosystem arrives, then one must use this method (or one like it) to construct something that is <:AbstractGraph.

Note

Requires Graphs to be loaded.

GraphDynamicalSystems.Constructors.vertex_function_to_edgelist — Method

vertex_function_to_edgelist(vertex_fn_pairs)
vertex_function_to_edgelist(vertex, fn, all_vertices)

Create a list of edges based on the vertices that appear in fn.

If a vertex X has a function X + Y and all_vertices = [X, Y, ...], then return edges X => X and X => Y.

Note

Requires AbstractTrees to be loaded.