Reference

Contents

Reference

Contents

Index

Index

GraphDynamicalSystems.BooleanNetworks

GraphDynamicalSystems.BooleanNetworks.BooleanNetwork

GraphDynamicalSystems.QN

GraphDynamicalSystems.QualitativeNetwork

DynamicalSystemsBase.get_state

DynamicalSystemsBase.get_state

DynamicalSystemsBase.set_state!

GraphDynamicalSystems.BooleanNetworks.abn

GraphDynamicalSystems.BooleanNetworks.abn

GraphDynamicalSystems.BooleanNetworks.abn_step!

GraphDynamicalSystems.BooleanNetworks.sample_boolean_network

GraphDynamicalSystems.BooleanNetworks.truth_dict_from_state

GraphDynamicalSystems.BooleanNetworks.update_functions_to_network

GraphDynamicalSystems.aqn

GraphDynamicalSystems.aqn

GraphDynamicalSystems.async_qn_step!

GraphDynamicalSystems.build_qn_grammar

GraphDynamicalSystems.components

GraphDynamicalSystems.interpret

GraphDynamicalSystems.limit_change

GraphDynamicalSystems.max_level

GraphDynamicalSystems.sample_qualitative_network

GraphDynamicalSystems.sample_qualitative_network

GraphDynamicalSystems.target_functions

GraphDynamicalSystems.update_functions_to_network

GraphDynamicalSystems.QN — Type

Shorthand for QualitativeNetwork.

GraphDynamicalSystems.QualitativeNetwork — Type

struct QualitativeNetwork

A qualitative network model as described in "Qualitative networks: a symbolic approach to analyze biological signaling networks".

graph: Graph containing the topology and target functions of the network

state: State of the network

N: The maximum activation level/state value of any component

Systems that include the model semantics wrap around this struct with an ArbitrarySteppable from DynamicalSystems. See aqn for an example.

DynamicalSystemsBase.get_state — Method

get_state(qn::QualitativeNetwork, component::Symbol) -> Any

DynamicalSystemsBase.get_state — Method

get_state(qn::QualitativeNetwork) -> AbstractVector{Int64}

DynamicalSystemsBase.set_state! — Method

set_state!(
    qn::QualitativeNetwork,
    component::Symbol,
    value::Integer
) -> Integer

GraphDynamicalSystems.aqn — Method

aqn(
    network::MetaGraphsNext.MetaGraph,
    initial_state::AbstractVector{Int64},
    max_level::Int64
) -> DynamicalSystemsBase.ArbitrarySteppable{U, QualitativeNetwork, typeof(GraphDynamicalSystems.async_qn_step!), typeof(GraphDynamicalSystems.reset_model!), typeof(GraphDynamicalSystems.extract_state), typeof(GraphDynamicalSystems.extract_parameters), P, DynamicalSystemsBase.var"#21#23"{typeof(GraphDynamicalSystems.reset_model!)}} where {U<:AbstractVector{Int64}, P<:MetaGraphsNext.MetaGraph}

Construct an asynchronous QualitativeNetwork system using the async_qn_step! as a step function.

GraphDynamicalSystems.aqn — Method

aqn(
    network::MetaGraphsNext.MetaGraph,
    max_level::Int64
) -> DynamicalSystemsBase.ArbitrarySteppable{U, QualitativeNetwork, typeof(GraphDynamicalSystems.async_qn_step!), typeof(GraphDynamicalSystems.reset_model!), typeof(GraphDynamicalSystems.extract_state), typeof(GraphDynamicalSystems.extract_parameters), P, DynamicalSystemsBase.var"#21#23"{typeof(GraphDynamicalSystems.reset_model!)}} where {U<:AbstractVector{Int64}, P<:MetaGraphsNext.MetaGraph}

GraphDynamicalSystems.async_qn_step! — Method

async_qn_step!(qn::QualitativeNetwork) -> Integer

GraphDynamicalSystems.build_qn_grammar — Method

build_qn_grammar(
    entity_names::AbstractVector,
    constants::AbstractVector{<:Integer}
) -> Any

Builds a grammar based on the base QN grammar adding entity_names and constants to the grammar.

Four constraints are currently included

removing symmetry due to commutativity of +/*/min/max

forbidding same arguments of two argument functions

forbidding trivial inputs (consts and entity values) to floor/ceil

forbidding ceil(floor(_)) and floor(ceil(_))

GraphDynamicalSystems.components — Method

components(qn::QualitativeNetwork) -> Any

GraphDynamicalSystems.interpret — Method

interpret(
    e::Union{Int64, Expr, Symbol},
    qn::QualitativeNetwork
) -> Any

Interpret target functions from a QualitativeNetwork.

GraphDynamicalSystems.limit_change — Method

limit_change(prev_value, next_value, N::Int64) -> Any

GraphDynamicalSystems.max_level — Method

max_level(qn::QualitativeNetwork) -> Int64

GraphDynamicalSystems.sample_qualitative_network — Method

sample_qualitative_network(
    entities::AbstractVector{Symbol},
    max_eq_depth::Int64
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, _A, _B, Nothing, GraphData, MetaGraphsNext.var"#11#13", Float64} where {_A, _B, GraphData<:HerbCore.AbstractGrammar}

GraphDynamicalSystems.sample_qualitative_network — Method

sample_qualitative_network(
    size::Int64,
    max_eq_depth::Int64
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, _A, _B, Nothing, GraphData, MetaGraphsNext.var"#11#13", Float64} where {_A, _B, GraphData<:HerbCore.AbstractGrammar}

GraphDynamicalSystems.target_functions — Method

target_functions(qn::QualitativeNetwork) -> Dict

GraphDynamicalSystems.update_functions_to_network — Method

update_functions_to_network(
    update_functions::AbstractDict{Symbol, Union{Int64, Expr, Symbol}},
    grammar::HerbCore.AbstractGrammar
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, _A, _B, Nothing, GraphData, MetaGraphsNext.var"#11#13", Float64} where {_A, _B, GraphData<:HerbCore.AbstractGrammar}

GraphDynamicalSystems.BooleanNetworks — Module

GraphDynamicalSystems.BooleanNetworks.BooleanNetwork — Type

mutable struct BooleanNetwork

graph: The structure and update functions of the network

state: The state of the network

GraphDynamicalSystems.BooleanNetworks.abn — Method

abn(
    network::MetaGraphsNext.MetaGraph,
    initial_state::AbstractVector{Int64}
) -> DynamicalSystemsBase.ArbitrarySteppable{U, GraphDynamicalSystems.BooleanNetworks.BooleanNetwork, typeof(GraphDynamicalSystems.BooleanNetworks.abn_step!), typeof(GraphDynamicalSystems.BooleanNetworks.reset_model!), typeof(GraphDynamicalSystems.BooleanNetworks.extract_state), typeof(GraphDynamicalSystems.BooleanNetworks.extract_parameters), P, DynamicalSystemsBase.var"#21#23"{typeof(GraphDynamicalSystems.BooleanNetworks.reset_model!)}} where {U<:AbstractVector{Int64}, P<:MetaGraphsNext.MetaGraph}

Create an asynchronous Boolean network (ABN) with the given network, initial_state. The update step for an asynchronous Boolean network is to choose a random node and update its state, one node at a time. The random choice is uniform over all nodes. See abn_step!.

GraphDynamicalSystems.BooleanNetworks.abn — Method

abn(
    network::MetaGraphsNext.MetaGraph;
    seed
) -> DynamicalSystemsBase.ArbitrarySteppable{U, GraphDynamicalSystems.BooleanNetworks.BooleanNetwork, typeof(GraphDynamicalSystems.BooleanNetworks.abn_step!), typeof(GraphDynamicalSystems.BooleanNetworks.reset_model!), typeof(GraphDynamicalSystems.BooleanNetworks.extract_state), typeof(GraphDynamicalSystems.BooleanNetworks.extract_parameters), P, DynamicalSystemsBase.var"#21#23"{typeof(GraphDynamicalSystems.BooleanNetworks.reset_model!)}} where {U<:AbstractVector{Int64}, P<:MetaGraphsNext.MetaGraph}

Create an asynchronous Boolean network with a random initial state.

GraphDynamicalSystems.BooleanNetworks.abn_step! — Method

abn_step!(
    model::GraphDynamicalSystems.BooleanNetworks.BooleanNetwork
) -> Any

Update step for an asynchronous Boolean network (see abn). At each step, It selects a node in the network and applies its update function, updating the state for the selected node with the update function's output.

GraphDynamicalSystems.BooleanNetworks.sample_boolean_network — Function

sample_boolean_network(
    n::Int64;
    ...
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Int64, SoleLogics.Formula, Nothing, Nothing, MetaGraphsNext.var"#11#13", Float64}
sample_boolean_network(
    n::Int64,
    depth::Int64;
    ...
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Int64, SoleLogics.Formula, Nothing, Nothing, MetaGraphsNext.var"#11#13", Float64}
sample_boolean_network(
    n::Int64,
    depth::Int64,
    seed::Int64;
    tactic
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Int64, SoleLogics.Formula, Nothing, Nothing, MetaGraphsNext.var"#11#13", Float64}

Return a random Boolean network with n nodes.

GraphDynamicalSystems.BooleanNetworks.truth_dict_from_state — Method

truth_dict_from_state(
    state::AbstractVector{Int64},
    labels::AbstractVector
) -> SoleLogics.TruthDict{Dict{A, T}} where {A<:SoleLogics.Atom, T<:SoleLogics.Truth}

GraphDynamicalSystems.BooleanNetworks.update_functions_to_network — Method

update_functions_to_network(
    update_functions::AbstractVector{<:SoleLogics.Formula}
) -> MetaGraphsNext.MetaGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Int64, SoleLogics.Formula, Nothing, Nothing, MetaGraphsNext.var"#11#13", Float64}